requirements.txt
python-telegram-bot==12.6.1
requests==2.23.0